// frontend/src/core/types.ts
export interface Message {
    id: number;
    username: string;
    content: string;
    is_read: boolean;
    is_edited: boolean;
    timestamp: string;
    profile_picture?: string;
    reply_to?: Message;
}
// ...
export interface User {
    id: number;
    created_at: string;
    last_seen: string;
    online: boolean;
    username: string;
    admin?: boolean;
    bio?: string;
}